edinet-wrapper/pyproject.toml
[project]
name = "edinet-wrapper"
version = "0.1.0"
description = "EDINET データ取得・パース用パッケージ（サブモジュール非依存）"
requires-python = ">= 3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
    "requests>=2.32.3",
    "polars>=1.26.0",
    "loguru>=0.7.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/edinet_wrapper"]
